Implemented per-session and per-hostgroup backend server stickyness functionality
The feature is off by default and can be enabled through a query annotation hint: sticky_backend.

The setting is session-wide and can be changed with any query.

The setting starts taking effect with the current query being executed, meaning that if
stickyness was OFF previously and the query is enabling it, a load-balanced backend server will be selected for this query,
and the following queries will be executed against the same server; if stickyness was previously ON and the current query disables it,
then the current query will be executed against a load-balanced server, so will all following queries.

For example the following query sets the stickyness setting to level 1 for the rest of the session:
SELECT 1 /* sticky_backend=1 */;

The levels are:
0 - disabled
1 - sticky server selection, fallback to load-balancing to another server in case the sticky server becomes unavailable
2 - strictly sticky - in case the previously selected server becomes unavailable, the query retries until time-out and then errors out